requirements.txt
fastapi==0.115.13
iniconfig==2.1.0
jose==1.0.0
packaging==25.0
pydantic==2.11.7
PyJWT==2.10.1
pytest==8.4.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
starlette==0.46.2
uvicorn==0.35.0
